meta: refactor some tests
